+++ CHANGELOG.md
@@ -1,3 +1,36 @@
+## 0.0.49
+
+- `ASTTypeDouble`:
+  - `acceptsType`: now also accepts an `ASTTypeInt`.
+- `wasm_generator.dart`:
+  - Fix `isBits64`.
+- Improve Wasm test coverage.
+
+## 0.0.48
+
+- `ASTNode`: 
+  - Now is a mixin.
+  - `getNodeIdentifier`: added optional parameter `requester`.
+  - Added `children` and `descendantChildren`.
+- `ASTFunctionDeclaration`:
+  - `getNodeIdentifier`: now can also resolve identifiers inside statements.
+ 
+- Wasm:
+  - Encode function names with UTF-8.
+  - `Wasm64`: added `i64WrapToi32`.
+- `WasmContext`:
+  - Added `returns` state. 
+- `ApolloGeneratorWasm`:
+  - Added `_autoConvertStackTypes`.
+  - `generateASTStatementReturnValue` and `generateASTStatementReturnVariable`:
+    - Auto cast returning types. 
+
+- Dart CI: added job `test_exe`.
+
+## 0.0.47
+
+- Improve variable type resolution while compiling to Wasm.
+
 ## 0.0.46
 
 - Improve type resolution of `ASTTypeVar`.